requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
pandas>=2.0.0
plotly>=5.18.0
pyyaml>=6.0
reportlab>=4.1.0
pillow>=10.0.0
